Fixed function testing that was largely failing due to missing operator functions

+++ server/functions/length.go
@@ -27,12 +27,12 @@ func init() {
 // length_varchar represents the PostgreSQL function of the same name, taking the same parameters.
 var length_varchar = framework.Function1{
 	Name:       "length",
-	Return:     pgtypes.Int64,
+	Return:     pgtypes.Int32,
 	Parameters: []pgtypes.DoltgresType{pgtypes.VarCharMax},
 	Callable: func(ctx framework.Context, val1 any) (any, error) {
 		if val1 == nil {
 			return nil, nil
 		}
-		return int64(len([]rune(val1.(string)))), nil
+		return int32(len([]rune(val1.(string)))), nil
 	},
 }
